requirements.txt
gradio
transformers>=4.45.0
torch
accelerate
bitsandbytes
python-dotenv
beautifulsoup4
yt-dlp
google-api-python-client
google-auth-httplib2
google-auth-oauthlib